pyproject.toml
[tool.poetry]
name = "test_django"
version = "0.1.0"
description = ""
authors = ["ma4akk <ma4akk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
python-decouple = "^3.6"
psycopg2 = "^2.9.3"
django-mptt = "^0.13.4"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-debug-toolbar = "^3.2.4"
isort = "^5.10.1"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
